adding cart ancd cart context functional component
