+++ src/App.jsx
@@ -4,6 +4,7 @@ import "./App.css";
 import { CueForm } from "./components/cue-form";
 import { Table } from "./components/table";
 import { handleSubmit } from "./services/handleSubmit";
+import { CartProvider } from "./components/cart-context";
 
 class App extends Component {
     constructor(props) {
@@ -13,15 +14,17 @@ class App extends Component {
 
     render = () => {
         return (
-            <main>
-                <HelmetProvider>
-                    <Helmet>
-                        <title>Smac-Corpus</title>
-                    </Helmet>
-                </HelmetProvider>
-                <CueForm onSubmit={this.handleSubmit} />
-                <Table data={this.state && this.state.data} />
-            </main>
+            <CartProvider>
+                <main>
+                    <HelmetProvider>
+                        <Helmet>
+                            <title>Smac-Corpus</title>
+                        </Helmet>
+                    </HelmetProvider>
+                    <CueForm onSubmit={this.handleSubmit} />
+                    <Table data={this.state && this.state.data} />
+                </main>
+            </CartProvider>
         );
     };
 }
